% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abort_bad_argument}
\alias{abort_bad_argument}
\title{Generate helpful error messages for invalid function inputs}
\usage{
abort_bad_argument(arg, must, not = NULL)
}
\arguments{
\item{arg}{the invalid function argument}

\item{must}{what a valid argument must be (e.g "numeric", or "all_models")}

\item{not}{what a valid argument cannot be (e.g. "character" or "adlkfm_modelz")}
}
\value{
an error. prints the custom message to the console.
}
\description{
This function appears in Advanced R Chapter 8.5.2 Signalling
\url{https://adv-r.hadley.nz/conditions.html#signalling}
It was presumably written by Hadley Wickham.
Generates a helpful error message and passes it to rlang::abort. Does not determine argument validity. Use this function inside a statement which determines that an input is invalid. Is adapted with extremely minor modifications from
}
