% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_model_preds.R
\name{add_component_predictions}
\alias{add_component_predictions}
\title{Add predictions for individual model components}
\usage{
add_component_predictions(by_var, ...)
}
\arguments{
\item{by_var}{A nested tibble, as output by add_nls()}

\item{...}{additional arguments to be passed to downstream functions.}
}
\value{
The input nested df, with the columns tidied and data.
}
\description{
I honestly don't exactly re
}
