% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_fits.R
\name{add_component_preds}
\alias{add_component_preds}
\title{Add predictions for individual model components}
\usage{
add_component_preds(df)
}
\arguments{
\item{df}{nested tibble}
}
\value{
nested tibble with individual components added to the nested dataframes

the df, with component predictions added
}
\description{
Add predictions for individual model components
}
