% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_tmas.R
\name{add_component_tmas}
\alias{add_component_tmas}
\title{Add tmas for components to a nested tibble}
\usage{
add_component_tmas(df)
}
\arguments{
\item{df}{nested tibble}
}
\value{
nested tibble with column contaiing tmas from components
}
\description{
Add tmas for components to a nested tibble
}
