% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drfu_tmas.R
\name{add_drfu_tmas}
\alias{add_drfu_tmas}
\title{Add drfu tmas as a column to nested dsf data}
\usage{
add_drfu_tmas(by_var, .data_col = "data", ...)
}
\arguments{
\item{by_var}{a nested tibble. one unique dataset per row, with data nested as a tibble. each tibble contains temperature, value, and drfu data}

\item{.data_col}{name of the column containing the nested data tibbles}

\item{...}{arguments passed to get_drfu_tmas, such as .x_vec (name of temperature column, defaults to "Temperature), .y_vec (name of dRFU column, defaults to "drfu_norm"), .precision (precision to determine tmas to, defalts to 0.1)}
}
\value{
the by_var tibble, with a new column "drfu_tma"
}
\description{
Add drfu tmas as a column to nested dsf data
}
