% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_estimates.R
\name{add_estimates}
\alias{add_estimates}
\title{Add estimates for key feautres in data}
\usage{
add_estimates(by_var, ...)
}
\arguments{
\item{by_var}{A nested tibble, as output by tidy_for_tmas. Contains at least two columns; .var, a character column containing unique identifiers for each dataset, by which by_var is nested, and data, a nested tibble containing numeric columns corresponding to normalized temperature, value, and the first derivative of the normalized value. Default names are "Temperature_norm", "value_norm", and "drfu_norm", corresponding to the names output by tidy_for_tmas. However, alternative column names may used by supplying them in the ... argument. These names are passed to get_estimates.}

\item{...}{additional named argument, which will be passed to the helper functions get_estimates, estimate, and tidy_estimates}
}
\value{
The input nested df, with an additional column named "est", containing a nested tibble of starting estimates, as created by tidy_estimates, for each row.
}
\description{
A wrapper function to apply get_estimates across a tibble of nested dsf data, and append a column of estimates to that input tibble, named "est". "est" contains the output of get_estimates for each variable.
}
\seealso{
\code{\link{tidy_for_tmas}}

\code{\link{get_estimates}}
}
