% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dye_screen.R
\name{add_facet_grid}
\alias{add_facet_grid}
\title{Add specific facet grids to to a ggplot2 object}
\usage{
add_facet_grid(row_var, col_var)
}
\arguments{
\item{row_var}{unquoted name of a single variable mapped to facet rows.}

\item{col_var}{unquoted name of a single variable mapped to facet columns}
}
\value{
a ggplot2 object, with grid facets added.
}
\description{
has scales = "free"
}
