% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dye_screen.R
\name{add_facet_wrap_10}
\alias{add_facet_wrap_10}
\title{Add specific facet wrapping to to a ggplot2 object}
\usage{
add_facet_wrap_10(...)
}
\arguments{
\item{...}{unquoted names of the variables by which to facet.}
}
\value{
a ggplot2 object, with facets added.
}
\description{
Has ncol = 10 and scales = "free"
}
