% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_model_preds.R
\name{add_model_preds}
\alias{add_model_preds}
\title{Title}
\usage{
add_model_preds(by_var, ...)
}
\arguments{
\item{by_var}{A nested tibble, as output by add_nls()}

\item{...}{additional named arguments, passed to downstream functions. Also serves the purpose of allowing irrelevant arguments passed from upstream functions using ... to be ignored.}
}
\value{
The input nested df, with the data column updated to include predictions for the individual sigmoids from the model fits.
}
\description{
Title
}
