% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_model_tmas.R
\name{add_model_tmas}
\alias{add_model_tmas}
\title{Add dRFU tmas to nested dataframe}
\usage{
add_model_tmas(by_var, ...)
}
\arguments{
\item{by_var}{nested tibble, as output by add_model_preds()}

\item{...}{additional named arguments, passed to downstream functions. Also serves the purpose of allowing irrelevant arguments passed from upstream functions using ... to be ignored.}
}
\value{
The input nested df, with an additioan column, model_tma,as calculated using the maximum of the dRFU for each reconstructed model component.
}
\description{
Add dRFU tmas to nested dataframe
}
