% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nls.R
\name{add_nls}
\alias{add_nls}
\title{Add model fits to nested data}
\usage{
add_nls(by_var, formula, lower_bound, control_list = list(maxiter = 500), ...)
}
\arguments{
\item{by_var}{A nested tibble, as output by add_start_pars().}

\item{formula}{formula for the model to be fit}

\item{lower_bound}{named numeric vector containing lower bounds, in order, for the input parameters.}

\item{control_list}{a named list, passed directly to nlslM, to set controls for the fitting.}

\item{...}{additional named arguments, which will be passed to the helper function get_nls. Also serves the purpose of allowing irrelevant}
}
\value{
The input nested df, with an additional column named "model", containing the input formula fitted to the data. Is either an s3 model class, or s3 faild model class.
}
\description{
Add model fits to nested data
}
