% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_model_preds.R
\name{add_standard_preds}
\alias{add_standard_preds}
\title{Add standard predictions from modelr to data.}
\usage{
add_standard_preds(by_var, ...)
}
\arguments{
\item{by_var}{A nested tibble, as output by add_nls()}

\item{...}{additional arguments to be passed to downstream functions}
}
\value{
The data tibble, with residuals and predictions added, as calculated by modelr::add_residuals and modelr::add_predictions.
}
\description{
Add standard predictions from modelr to data.
}
