% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_start_pars.R
\name{add_start_pars}
\alias{add_start_pars}
\title{add starting parameters}
\usage{
add_start_pars(by_var, ...)
}
\arguments{
\item{by_var}{do this}

\item{...}{do this}
}
\value{
do this
}
\description{
add starting parameters
}
