% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_fits.R
\name{annotate_tidied_components}
\alias{annotate_tidied_components}
\title{Annotate the components from the model parameter outputs}
\usage{
annotate_tidied_components(tidied)
}
\arguments{
\item{tidied}{a tibble, the tidied model output}
}
\value{
the tidied model output, now with additional annotation columns

components, annotated extra columns
}
\description{
Annotate the components from the model parameter outputs
}
