% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dye_screen_figs.R
\name{blank_label_plot}
\alias{blank_label_plot}
\title{Create blank ggplot with a label}
\usage{
blank_label_plot(.text, .text_preamble = "No dyes with assignment: ", ...)
}
\arguments{
\item{.text}{a character string, appended after .text_preamble, and printed to the blank plot. Typically used to pass assignment to the label, e.g "hit"}

\item{.text_preamble}{a character string, which appears directly before .text. Defaults to No dyes with assignment: "}

\item{...}{additional named arguments, passed to \code{geom_text()}}
}
\value{
a ggplot2 object containing a blank plot with \code{theme_void()}, with the specified label. Will appear just as text when displayed.
}
\description{
The \code{blank_label_plot()} function is used to generate a blank plot which prints a given string of text. Used in the \code{save_*_screen()} family functions, to handle cases where one of the expected assignments (e.g. "hit", "sensitive", "none") have been given to no dyes.
}
