% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_estimates.R
\name{cluster_estimates}
\alias{cluster_estimates}
\title{Cluster both estimated transitions}
\usage{
cluster_estimates(est_df, .est_row_col = "est_row", min_dist = 4, ...)
}
\arguments{
\item{est_df}{a tibble of estimated transitions, as output by tidy_estimates}

\item{.est_row_col}{a string, providing the name of the numeric column passed to hclust. Defaults to "est_row", the measurment number of the estimated transtion.}

\item{min_dist}{the minimum distance between two estimates. Defaults to 4. For a dataset with one measurement per degree, this corresponds to a minimum inter-estimate distance of 4 degrees.}

\item{...}{permit the presence additional parameters, which may be passed via ... from upstream functions, but are ignored here.}
}
\value{
the input tibble, with identical format, but the input .est_row_col overwritten by the cluster number to which the estimate belongs.
}
\description{
Groups estimated transtions, without distinguishing between major and minor, using hclust. Sorts estimations into clusters, and overwrites the input est_row with the cluster number to which the estimate was assigned. Individual clusters are defined by a minimum inter-cluster distance. With default usage, clusters the estimates using the row/measurement number of the estimates (corresponding to temperature), with minimum inter-measurment distance of 4.
}
\details{
I don't recall how carefully I considered the downstream consequences of over-writing the given row. If the clustering fails, is all of this informaiton lost? It might be good to specify a safer behavior in the hclust_points function.
}
\seealso{
\code{\link{hclust_points}}
}
