% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dye_screen_figs.R
\name{convert_heights}
\alias{convert_heights}
\title{Not exported: helper function for the \code{save_stacked_plots()} function.}
\usage{
convert_heights(
  .paneled_by,
  .raw_panel_height = 0.4,
  facet_type = "wrap",
  .ncol_wide = 10,
  .wrap_margin_ratio = 190/110,
  .grid_margin_ratio = 190/110,
  .title_height_add = 1
)
}
\arguments{
\item{.paneled_by}{a list or vector containing the variable by which the plot was faceted. used only for it's length.}

\item{.raw_panel_height}{the fored height of the panel. defaults to 0.4}

\item{facet_type}{the type of faceting applied to the plot, "wrap" or "grid", because this number is different for wraps anada grids.}

\item{.ncol_wide}{the number of columns in the facet, used to calculate the number of rows in the facet.}

\item{.wrap_margin_ratio}{a number, giving the ratio between the height of a full panel and the height of just it's forced plot area, for wrapped facets. Defaults to 190/110, but this may be a bit too large for standard applications.}

\item{.grid_margin_ratio}{a number, giving the ratio between the height of a full panel and the height of just it's forced plot area, for grid facets. Defaults to 190/110, but this may be a bit too large for standard applications.}

\item{.title_height_add}{a number, giving a constant amount to add to each plot to accomodate single-appearance items like titles and legends.}
}
\value{
a number, giving the estimated height of a faceted plot created based on the variable defined in paneled-by
}
\description{
This is a brute-force way of getting around challenges of autoamtically determining the height of a full ggplot object. While we can set the panel heights precisely, I haven't figured out a good way to set all additional obejct sizes to allow perfect computation of output plot height (e.g. axis text, axis ticks, margins...). This is a simple arithmetic operation, but i pulled it into a helper function to allow more robust tuning of this converstion, and to simplify the code for the functions which use it.
}
\details{
Note that there is at least one function get_dims from the package infotroph/DeLuciatoR, which does this calculation, but function relied on unstable side effects of certain functions. I didn't use it, because I don't want to take a dependency on something like that.
}
