% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model_pars.R
\name{estimate_transitions}
\alias{estimate_transitions}
\title{Estimate locations of transitions in data}
\usage{
estimate_transitions(
  values,
  norm_raw_y,
  norm_raw_x,
  transition = "major",
  low_temp_margin = 5,
  high_temp_margin = 95
)
}
\arguments{
\item{values}{numeric vector containing a single raw dataset}

\item{norm_raw_y}{numeric vector containing normalized raw rfu data}

\item{norm_raw_x}{numeric vector containing normalized temperature data}

\item{transition}{which transition to detect. options: "major" (peaks in drfu), "minor" (valleys in ddrfu), "initial" (the starting rfu value)}

\item{low_temp_margin}{do not take any values in the first x measurements  (e.g. don't take a peak if it's the first value)}

\item{high_temp_margin}{do not take any values after the first x measurement (e.g. don't take a peak if it's the last value)}
}
\value{
tibble containing potential transition locations
}
\description{
Estimate locations of transitions in data
}
