% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_tmas.R
\name{extract_component_tma}
\alias{extract_component_tma}
\title{Extract tmas from each component using max drfu method}
\usage{
extract_component_tma(
  data,
  component_names = c("sigmoid_1", "sigmoid_2", "sigmoid_3")
)
}
\arguments{
\item{data}{dataframe containing tma components}

\item{component_names}{names of the components to extract tmas from}
}
\value{
tibble containing tmas and their numbering (e.g. tma1, tma2...)
}
\description{
Extract tmas from each component using max drfu method
}
