% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_fits.R
\name{fit_all}
\alias{fit_all}
\title{Fit all supplied datasets to a single requested model}
\usage{
fit_all(
  by_variable,
  which_pars,
  formula,
  lower_bound,
  drop_raw = TRUE,
  temperature_col = c("Temperature"),
  keep_data_cols = c("resid", "pred"),
  ...
)
}
\arguments{
\item{by_variable}{nested tibble}

\item{which_pars}{names of the parameters in the model}

\item{formula}{formula to fit}

\item{lower_bound}{lower bound for model parameters}

\item{drop_raw}{drop the raw data from the results? defaults to TRUE}

\item{temperature_col}{name of column containing temperature data. defaults to c("Temperature")}

\item{keep_data_cols}{names of columns to retain in the fit results. defaults to c("resid", "pred")}

\item{...}{more arguments, passed to the fit}
}
\value{
a dataframe containing nested rows with the fitted model, and tidied, glanced, predicted, and residual values
}
\description{
Fit all supplied datasets to a single requested model
}
