% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_fits.R
\name{fit_nls}
\alias{fit_nls}
\title{Fit dsf data to a model}
\usage{
fit_nls(
  df,
  start_pars,
  formula,
  lower_bound,
  control_list = list(maxiter = 500)
)
}
\arguments{
\item{df}{tibble containing data to fit}

\item{start_pars}{list of starting parameters, passed to nlsLM}

\item{formula}{model formula to fit, passed to nlsLM}

\item{lower_bound}{a vector of ower bounds of fit, passed to nlsLM}

\item{control_list}{list of controls for nlsLM. Defaults to list(maxiter = 500)}
}
\value{
fitted model, or, if it fails to converge, NA
}
\description{
Fit dsf data to a model
}
