% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dye_screen.R
\name{force_panel_sizing}
\alias{force_panel_sizing}
\title{Set specific panel sizes}
\usage{
force_panel_sizing(
  .force_height = 0.4,
  .force_width = 0.4 * 1.618,
  .units = "in",
  ...
)
}
\arguments{
\item{.force_height}{height of each panel in a facet. Defaults to 0.4}

\item{.force_width}{width of each panel in a facet. Defaults to 0.4 * 1.618 (golden ratio)}

\item{.units}{units of the given values. Defaults to "in"}

\item{...}{additoinal arguments, to be tolerated from upstream functions but do nothing here.}
}
\value{
a ggplot2 object, with facets of specific sizes.
}
\description{
Set specific panel sizes
}
