% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_fits.R
\name{get_component_exprs}
\alias{get_component_exprs}
\title{Get expressions for each omponent}
\usage{
get_component_exprs(component_type)
}
\arguments{
\item{component_type}{the type of component}
}
\value{
quoted expressioned for each model component
}
\description{
Get expressions for each omponent
}
