% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_model_preds.R
\name{get_component_predictions}
\alias{get_component_predictions}
\title{Get component predictions}
\usage{
get_component_predictions(
  tidied,
  data,
  temp_col = "Temperature_norm",
  rescale_temp_col = TRUE,
  ...
)
}
\arguments{
\item{tidied}{a tibble resulting from annotate_tidied_components}

\item{data}{a tibble of dsf data, resulting from drop_and_pivot, to which component predictions will be added.}

\item{temp_col}{a string, containing the name of the column in the dataframe supplied to the data argument, whch contains the x data used to fit the model. Defaults to Temperature_norm. If the original model was fit to Temperature_norm, this must also be Temperature_norm--it cannot be "Temperature"}

\item{rescale_temp_col}{boolean, that, if true, triggers the rescaling of the input Temperautre column to a 0 to 1 range.}

\item{...}{additional parameters passed to downstream functions or ignored.}
}
\value{
the data tibble, with additional rows corresponding to the reconstructed component.
}
\description{
Get component predictions
}
