% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drfu_tmas.R
\name{get_drfu_tmas}
\alias{get_drfu_tmas}
\title{Calculate apparent melting temperatures using first derivative}
\usage{
get_drfu_tmas(
  data,
  .x_vec = "Temperature",
  .y_vec = "drfu_norm",
  .precision = 0.1
)
}
\arguments{
\item{data}{a tibble for a single dataset containing temperature and dRFU (first derivative) as columns}

\item{.x_vec}{name of the column containing temperature data}

\item{.y_vec}{name of the column containing dRFU data}

\item{.precision}{precision with which to determine tma}
}
\value{
a number; the interpolated maximum of the dRFU data
}
\description{
Calculate apparent melting temperatures using first derivative
}
