% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_estimates.R
\name{get_estimates}
\alias{get_estimates}
\title{Estimate key values from in a single trace of DSF data.}
\usage{
get_estimates(
  df,
  .est_peak_col = "drfu_norm",
  .est_init_col = "value_norm",
  .norm_raw_x = "Temperature_norm",
  .norm_raw_y = "value_norm",
  ...
)
}
\arguments{
\item{df}{A tibble of DSF data containing numeric columns corresponding to temperature, normalized RFU, and first derivative of normalized RFU.}

\item{.est_peak_col}{name of the column, as a string, containing data in which to look for major peaks. Defaults to "drfu_norm".}

\item{.est_init_col}{name of the column, as a string, containing data in which to look for inital RFU values.  Defaults to "value_norm".}

\item{.norm_raw_x}{name of the column, as a string, containing normalizd temperature data. Defaults to "Temperature_norm".}

\item{.norm_raw_y}{name of the column, as a string, containing normalizd raw RFU data. Defaults to "value_norm".}

\item{...}{addional named arguments, passed to estimate, rank_estimates, and tidy_estimates}
}
\value{
a tibble, containing columns val_at_est (the value of the data used to define the estimate at the estiamted point), est_val (value of the actual estimate itself, which is passed to get_start_pars), est_raw_x (raw x value at the estimated point), est_raw_y (raw y value at the estimated point), est_type (type of estimate, "transition" or "initial")
}
\description{
Estimate the values of key features in a single DSF dataset. Is also a helper function for add_estimates. Returns a tibble containing estimations of key values for each input dataset. Key values include the normalized temperatures at which transitions may occur in the data, and the normalized value of the initial readings. These values are used to generate starting parameter estimates for each dataset in add_start_pars. Estimates are included here (rather than just start pars) to enable users to extract their own starting parameters for their own models. Estimates for locations of transitions are rank-ordered by rank_estimates, to simplify passing to start parameters for downstream fitting.
}
\seealso{
\code{\link{add_estimates}}
}
