% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nls.R
\name{get_nls}
\alias{get_nls}
\title{Fit a given formula to single trace of DSF data}
\usage{
get_nls(
  df,
  start_pars,
  formula,
  lower_bound,
  control_list = list(maxiter = 500),
  ...
)
}
\arguments{
\item{df}{a tibble containing x and y data for a single trace of DSF data. In default use, this is Temperature_norm and value_norm.}

\item{start_pars}{named list of starting parameters for the supplied formula. Passed to nlsLM.}

\item{formula}{formula to fit. Passed to nlsLM.}

\item{lower_bound}{a named numeric vector, in identical parameter order to start_pars, containing lower bounds for the input functions.}

\item{control_list}{an optional named list of controls to pass to nlslM, including max interations.}

\item{...}{additional arguments, which, if inherited from upstream functions, can be ignored by this function.}
}
\value{
a fitted model, exactly as output by nlslM. Or, if the nlsLM operation failed, returns NA.
}
\description{
Fit a given formula to single trace of DSF data
}
