% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_start_pars.R
\name{get_start_pars}
\alias{get_start_pars}
\title{Select and format estimates for model starting parameters}
\usage{
get_start_pars(
  estimates,
  which_pars,
  pars_defaults = list(Asym1 = 1, xmid1 = 0.4, scal1 = 0.03, d1 = -0.5, Asym2 = 0.5,
    xmid2 = 0.6, scal2 = 0.03, d2 = -1, Asym3 = 0.3, xmid3 = 0.2, scal3 = 0.03, d3 =
    -1.5, id_d1 = 0.2, id_b1 = -5),
  par_order = c("Asym1", "xmid1", "scal1", "d1", "Asym2", "xmid2", "scal2", "d2",
    "Asym3", "xmid3", "scal3", "d3", "id_d1", "id_b1"),
  ...
)
}
\arguments{
\item{estimates}{a tibble, as generated by get_estimates, containing rank-ordered estimates for transitions and intial RFUs for a singe trace of DSF data}

\item{which_pars}{a character vector, containing the names of all parameters for which starting parameters should be returned.}

\item{pars_defaults}{a named list, containing default values to provide for starting parameters for all possible parameters used in dsfworld models.}

\item{par_order}{a character vector, containing the names of the parameters to be output, in the order in which they will be expected in the downstream nlsLM. WRITE A CHECK FOR THIS.}

\item{...}{additional arguments, inherited from upstream functions such as add_start_pars, which can be ignored by this function.}
}
\value{
a named list of numbers, with names corresponding to parameters names of a dsfworld model, and values corresponding to starting parameter estimates. List element appear in the order in which nslLM expects them. WRITE A CHECK FOR THIS
}
\description{
Select and format estimates for model starting parameters
}
