% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_estimates.R
\name{hclust_points}
\alias{hclust_points}
\title{Identify clusters within a numeric integer vector}
\usage{
hclust_points(points)
}
\arguments{
\item{points}{numeric vector of integrers containing values to be clustered.}
}
\value{
object as returned by hclust on a self-by-self distance matrix constructed from the supplied points vector.
}
\description{
A helper function to cluster_estimates. Given a numeric integer vector, constructs a self-by-self dissimilarity matrix. Applies hclust to this matrix to identify clusters within the input vector. Output is passed back to cluster_estimates.
}
\seealso{
\code{\link{cluster_estimates}}
}
