% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_data.R
\name{label_data}
\alias{label_data}
\title{Join raw data and plate layout tibbles into a single, labeled tibble}
\usage{
label_data(
  data,
  layout,
  join_by = "well",
  drop_empties = TRUE,
  .drop_from = protein,
  .drop_matches = c("Empty", "empty"),
  .make_numeric = c(),
  .make_character = c()
)
}
\arguments{
\item{data}{A tibble containg raw data to be joined with the layout tibble, and at least one column also present in the layout tibble, typically "well".}

\item{layout}{A tibble, containing at least one column also present in the data tibble, typically "well", which will be used to join with the data tibble. Additional columns in the layout tibble contain information associated with each well, typically experimental variables, which will also be joined with the data. Typically, the layout tibble is created by reading in plate layout file using the function read_plate_layout()}

\item{join_by}{A column shared between the data and layout, identical in name and related in content. Typically "well".}

\item{drop_empties}{TRUE or FALSE value, determining whether data from particular wells should be dropped from the output tibble. Defaults to TRUE.}

\item{.drop_from}{If drop_empties is set to TRUE, the unquoted name of the column which will be used to perform value-based filtering.}

\item{.drop_matches}{A vector containing all values which, if present in the .drop_from column, will trigger the removal of those wells from the output tibble.}

\item{.make_numeric}{A vector containing the names of any columns from the layout which, if not already numeric, should be coerced to numeric. Numeric-type columns are automatically detected using readr::parse_guess() in read_plate_layout, so this argument is necessary only if a user wants to override the decision made by readr::parse_guess().}

\item{.make_character}{A vector containing the names of any columns from the layout which, if not already of type character, should be coerced to character. Character-type columns are automatically detected using readr::parse_guess() in read_plate_layout, so this argument is necessary only if a user wants to override the decision made by readr::parse_guess().}
}
\value{
A tibble, containing all columns present in the input data and layout tibbles, joined into a single, labeled tibble.
}
\description{
label_data() joins two tibbles into one: raw data, and the experimental information provided by the plate layout. The single tibble which results, which contains both raw data and experimental conditions associated with each well, allows handling and visualiation of data based on experimental variables.
}
