% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_dye_screen.R
\name{label_dye_screen}
\alias{label_dye_screen}
\title{Label and clean dye screen data}
\usage{
label_dye_screen(
  .data,
  .layout,
  .type,
  .variable_col = "variable",
  .temp_col = "Temperature",
  .value_col = "value",
  .dye_col = "final_compound",
  .dye_conc_col = "final_concentration",
  .channel_col = "channel_f",
  .drop_layout_empties_from = "dye",
  .keep_additional_cols = "",
  ...
)
}
\arguments{
\item{.data}{a long-form tibble containing a single plate of raw DSF data from a dye screen. Must include columns containing information on each of the following things:
\itemize{
\item \strong{well}, or some other individual reaction-identifying information. This column is coerced to character-type in the output of this function.
\item \strong{Temperature}, giving the Temperature of each RFU measurement. This column is coerced to numeric-type in the output of this function.
\item \strong{Wavelength of measurement} A column giving information about the wavelengths at which measurements were made, e.g. "channel_f". This column must be a factor; it is recommended that the factor levels reflect the wavelengths measured by each channel. For example, "channel_f" may be a factor column, with levels of "FAM", "JOE", "TAMRA", "ROX", "Cy5", "Cy5.5". This column is added automatically by read_qtower, but may need to be generated manually for other types of inputs.

\item \strong{Raw RFU value}, giving the raw values measured. This column is coerced to numeric-type in the output of this function. A by-variable normalized version of this column is also present in the returned tibble.
}}

\item{.layout}{a long-form tibble containing plate layout information for the supplied raw data. Must include columns containing information on each of the following things:
#' \itemize{
\item \strong{well}, or some other individual reaction-identifying information which matches the identifying column in the supplied data.nThis column is coerced to character-type in the output of this function.
\item \strong{dye}, giving a unique identifying name for the dye that was present in a well. This column is coerced to character-type in the output of this function.
\item \strong{dye concentration}, a column giving the concentration (in uM) at which the dye was tested. This column is coerced to numeric-type in the output of this function. uM units are not precisely necessary, but the default name of this column for processing is "dye_conc_uM", so non-uM units may lead to misleading column labeling if default column names are used.
}}

\item{.type}{a string, giving either "protein" or "buffer". This value will be present in the output as a column entitled "type", with only this value below.}

\item{.variable_col}{a string, giving the name of the column containing unique identifying information for each trace of data. For DSF runs in which each well was measured in multiple channels, note that "variable" must comprise both the well and measurement channel to be unique, e.g. "A1_FAM", "A1_JOE", etc. Defaults to "variable".}

\item{.temp_col}{a string, giving the name of the column containing the temperature data. Defaults to "Temperature".}

\item{.value_col}{a string, giving the name of the column containing the raw RFU values. Defaults to "value".}

\item{.dye_col}{a string, giving the name of the column containing the unique dye names. Defaults to "final_compound".}

\item{.dye_conc_col}{a string, giving the name of the column containing the dye concentrations Defaults to "final_concentration".}

\item{.channel_col}{a string, giving the name of the column containing the channel information for the measurements. Defaults to "channel_f".}

\item{.drop_layout_empties_from}{a string, giving the name of the column  used to drop any NA-containing rows from the output data. Is passed to .drop_from in label_data, and defaults here to "dye"}

\item{.keep_additional_cols}{a character vector, giving names of any additional columns that should be retained in the final output beyond the default "variable", "Temperature", "value", "dye_conc_uM", "channel_f", "dye". This argument is useful in cases where additional information from the layout should be carried through the dye screen analysis process.}

\item{...}{additional arguments, passsed to label_data.}
}
\value{
a tibble, containing data for a labeled dye screen, of either protein or buffer. Columns are:
#' #' \itemize{
\item \strong{variable}, a character column holding unique identifiers for each trace.
\item \strong{Temperature},
\item \strong{value},
\item \strong{value_norm},
\item \strong{dye}, a character column specifying the dye tested in each well.
\item \strong{dye_conc_uM}, a character column specifying the concentration at which a dye was tested.
\item \strong{channel_f}, a character column specifying the channel in which the raw values were measured.
\item \strong{type}, a character column specifying the type of moleceule screened-- "protein" or "buffer". Each output may contain only one value for "type". For plates in which buffer and protein were both screened, layouts which mask the wells of the opposite type should be used, to create separate outputs for each type of data.

}
}
\description{
A helper function for tidy_dye_screen(). Ensures that individual runs contain the correct columns.
}
