% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dye_screen_model_plots.R
\name{plot_analyzed_hits}
\alias{plot_analyzed_hits}
\title{Plot}
\usage{
plot_analyzed_hits(
  tidied_for_plot,
  hits_by_channel,
  model_tmas,
  .dye_channel_col = "dye_channel_f",
  .plot_assignment = "hit",
  .var_col = ".var",
  ...
)
}
\arguments{
\item{tidied_for_plot}{a tibble of raw data, as returned by \code{tidy_for_tmas}}

\item{hits_by_channel}{a tibble, as expected by \code{pull_assigned}, containing assignments matching to values found in \code{tidied_fot_plot}, in the column specifided through the \code{.dye_channel_col} argument.}

\item{model_tmas}{a tibble, as returned by \code{extract_model_tmas}}

\item{.dye_channel_col}{a string, giving the name of the column in \code{tidied_for_plot} containing information on which hits were called, e.g. "dye_channel_f", which typically contains values in the format <dye_name>_<channel_name>, e.g. "T004_FAM".}

\item{.plot_assignment}{a string, passed to \code{pull_assigned}; all values with this assignment in \code{hits_by_channel} will appear in the output plot.}

\item{.var_col}{a string, giving the name of the column in both \code{tidy_for_tmas} and \code{by_var_fit} giving identical unqiue identifiers fo a given dataset. Defaults to ".var"}

\item{...}{additional arguments, passed to \code{plot_with_preds()}.}
}
\value{
a ggplot2 object, as returned by \code{plot_with_preds}, for a defined subset of the data, and with a plot title.
}
\description{
A wrapper function for \code{plot_with_preds}; \code{plot_analyzed_hits} uses an additional input tibble to filter the plotted data to include only hits, and adds custom plot titles. In a better version of this function, would likely be refactored alongside \code{plot_with_preds()} and \code{save_model_figure()}.
}
