% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dye_screen.R
\name{plot_subscreen}
\alias{plot_subscreen}
\title{Plot a dye screen for a subset of dyes, and either wrap or grid for facets.}
\usage{
plot_subscreen(
  tidied_screen,
  dyes,
  .facet_type = "wrap",
  plot_title = "",
  .dye_col = "dye",
  .channel_col = "channel_f",
  ...
)
}
\arguments{
\item{tidied_screen}{a tibble containing raw data from a dye screen, both protein and buffer, e.g. as output by tidy_dye_screen()}

\item{dyes}{a character vector, containing names of the dyes to be displayed in the output plot. Gets passed to filter. Must be present in the column with the name supplied in the .dye_col argument of this function (defaults to "dye")}

\item{.facet_type}{the type of faceting to apply. Options are "wrap" or "grid", which call the helper functions add_facet_wrap_10 or add_facet_grid, respectively.}

\item{plot_title}{a title to add to the plot.}

\item{.dye_col}{a string, containing the name of the column in user-supplied data 'tidied_screen', containing dye names which can be filtered using the supplied vector passed to the 'dyes' argument of this function.}

\item{.channel_col}{a string, containing the name of the column in user-supplied data 'tidied_screen', containing channels.}

\item{...}{additiona arguments, passed to force_panel_sizing}
}
\value{
a ggplot2 object containing dye screening data, complete with titles, facet sizes, and appropraite faceting.
}
\description{
A wrapper function for plot_dye_screen, meant to ease the creation of the standard plots (wrap by dye, and grid by dye and channel) created during a dye screen analysis workflow.
}
