% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_fits.R
\name{predict_component}
\alias{predict_component}
\title{Predict the values of a single component of a model}
\usage{
predict_component(df, x_vec = seq(from = 0, to = 1, length = 70), ...)
}
\arguments{
\item{df}{tibble with data}

\item{x_vec}{x vector to use for prediction.}

\item{...}{does nothing yet; to pass extra external pars eventually}
}
\value{
predicted components
}
\description{
Predict the values of a single component of a model
}
