% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dye_screen_figs.R
\name{pull_assigned}
\alias{pull_assigned}
\title{Not exported: helper function to the \code{save_stacked_plots()} function.}
\usage{
pull_assigned(hits, assignment, .dye_col = "dye", .assign_col = "assignment")
}
\arguments{
\item{hits}{the tibble containing hit assignments. See the function documentation for \code{save_stacked_plots()} for more information.}

\item{assignment}{a string, giving the assignment for which dyes should be pulled.}

\item{.dye_col}{a string, giving the name of the column in \code{'hits'} containing dye names.}

\item{.assign_col}{a string, giving the name of the column in \code{'hits'} containin assignments.}
}
\value{
a character vector containing the names of the dyes which matched to the provided assignment.
}
\description{
Not exported: helper function to the \code{save_stacked_plots()} function.
}
