% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_estimates.R
\name{rank_estimates}
\alias{rank_estimates}
\title{Rank estimated transitions by their predicted importance}
\usage{
rank_estimates(est_df, ...)
}
\arguments{
\item{est_df}{a tibble of estimated transitions, as output by tidy_estimates}

\item{...}{permit the presence additional parameters, which may be passed via ... from upstream functions, but are ignored here.}
}
\value{
do this
}
\description{
A helper function for get_estimates, which is also used inside add_estimates. Rank-orders estimated transitions by predicted importance, such that estiamted transitions are passed to model starting parameters in order of decreasin predicted importance. Handles the very common case in which more transitons are estimated from the input data than will be fit by the downstream model. The case where fewer estimates are identified than are fit in the downstream model is handled by default values provided to to in tidy_estimates.
}
\seealso{
\code{\link{get_estimates}}

\code{\link{add_estimates}}

\code{\link{tidy_estimates}}
}
