% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plate_layout.R
\name{read_plate_layout}
\alias{read_plate_layout}
\title{Read a plate layout file into a tibble}
\usage{
read_plate_layout(filepath)
}
\arguments{
\item{filepath}{A complete file path, pointing to the plate layout file}
}
\value{
Returns a tibble, mapping experimental variables to well positions. All outputs contain the columns: row, column, and well. Additionally, a single column is added for each user-defined variable in the plate layout file, from which one additional "condtion" column is created, which contains all experimental variables. If all experimental variables are defined in the layout, wells with identical entries in the "condition" column are technical replicates.
}
\description{
read_layout() reads a plate layout file (.csv, .txt, .xls, or .xlsx), and returns it as a formatted tibble. Variable types are guessed with readr::parse_guess(). The originally required "Type" column heading is now optional.
}
