% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qTower.R
\name{read_qtower}
\alias{read_qtower}
\title{Read raw data file as exported from a qTower}
\usage{
read_qtower(
  file_path,
  start_temp = 25,
  inc_temp = 1,
  channel_levels = c("FAM", "JOE", "TAMRA", "ROX", "Cy5", "Cy5.5", "SyproOrange")
)
}
\arguments{
\item{file_path}{A complete file path, pointing to the raw data (.csv) exported from a qTower}

\item{start_temp}{The temperature of the first read in the data file. Used to convert from cycle number to Temperature}

\item{inc_temp}{The increment by which the temperature increases with each read. Used to convert from cycle number to Temperature.}

\item{channel_levels}{A vector used to set the order of the channels measured in the loaded experiment. Defaults to FAM, JOE, TAMRA, ROX, Cy5 ,Cy5.5, SyproOrange. Channels not present in this list will be NA in the output tibble.}
}
\value{
A tibble containg all data, but no meta-data, present in the loaded csv file. Includes columns: variable (combines well and channel, to uniquely identify each trace), well, channel, channel_f (channels as a factor, ordered by increasing excitation wavelength), Temperature, value, and value_norm (each raw data trace, rescaled to a 0 to 1 range).
}
\description{
Read raw data file (.csv) directly as exported from a qTower 384G qPCR instrument. Creates a "channel_f" column, which contains the channels as a factor, with default levels of FAM, JOE, TAMRA, ROX, Cy5 ,Cy5.5, SyproOrange. Converts from cycle number to Temperature, with a default starting temperature of 25, and 1 C increase between measurements.
}
