% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dye_screen_figs.R
\name{save_dye_screen_figs}
\alias{save_dye_screen_figs}
\title{Save dye screen figures}
\usage{
save_dye_screen_figs(tidied_screen, plot_type, hits, ...)
}
\arguments{
\item{tidied_screen}{a tibble containing raw data from a dye screen, both protein and buffer, e.g. as output by the \code{tidy_dye_screen()} function.}

\item{plot_type}{type of figure to create. Options: "full_screen", "divided", "hits_by_channel", which generate and save figures created by the \code{save_full_screen()} function, the \code{save_stacked_screen()} function, and the \code{save_stacked_hits()} function respectively.}

\item{hits}{a tibble containing hit assignments for each dye. Must have columns "dye" and "assignment", described below:
\itemize{
\item \strong{"dye"}, a character column containing the names of all of the dyes tested in the experiment. In order for a dye to be displayed in this plot, its name in this dataframe must match its name in the tidied_screen dataframe exactly. Matches are case- and punctuation-sensitive. (e.g. a dye named "L045" in this tibble will match only to a dye named "L045" in the tidied_screen tibble; it will not match to "l045", "L45", "L-045", etc.).
\item \strong{"assignment"}, a character column containing the assignment of how a given dye behaved with screened protein. These assignments must be supplied by the user. Assignments can be determined either manually, or with assistance using the hit-calling tools in this package (in development as of August 2021). Possible assignments are: "hit", "sensitive", and "none", which correspond to the following judgements:
\itemize{
\item \strong{"hit"}, any dye which appears to detect the melting of the protein of interest, while having minimal temperature-sensitive fluorescence in the absence of protein.
\item \strong{"sensitive"}, a less stringent category than 'hit'. Useful for dyes where definition as a "hit" isn't quite warranted, but which demonstrating some fluorescent response to the protein (over buffer) which could be potentially interesting. Common reasons for a a dye to receive a "senstive" assignment detection of non-cannonical transitions, and/or detecting a cannonical transtion, but doing so with far weaker signal than dyes called as 'hits'. In particular, the "sensitive" assignment is useful for flagging dyes which may be interesting for follow-up later, or fleshing out SAR of the hit dyes for a given protein.
\item \strong{"none"}, an assignment given to dyes which showed no major difference between a DSF run with or without the protein, suggesting no fluorescent response to the protein of interest at any temperature.
}
}}

\item{...}{additional arguments passed to the \code{save_full_screen()} function, the \code{save_stacked_screen()} function, or the \code{save_stacked_hits()} function. Common use could be to specify a directory to save the output plot, via .save_path.}
}
\value{
if assigned to an object, returns a ggproto object. Also saves the figure.
}
\description{
The \code{save_dye_screen_figs()} function is used to quickly generate and save three of the four standard raw data figures in a dye screening workflow. See the 'Details' section below for more information on each of these plots.
}
\details{
The three plots which can be generating using \code{save_dye_screen_figs()} are:
\enumerate{
\item all raw data, with no assignments displayed. Used for calling hits, as created by the \code{save_full_screen()} function.
\item all raw data, with distinct sub-plots containing dyes given each assignment--hit, sensitive, or none--as created by the \code{save_stacked_screen()} function.
\item raw data for just dyes assigned to "hit" and "sensitive" categories, with data collected for each dye displayed in a separate row, and data for the raw data collected for that dye in each fluorescent channel displayed in a separate colummn, as created by the \code{save_stacked_hits()} function
\item Not accessed though this function. For reference, the fourth plot displays the hit and/or sensitive dyes, in specific channels, carried through model fitting and tma extraction, alongside the results of the model fitting and tma extraction.
}
}
