% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dye_screen_figs.R
\name{save_full_screen}
\alias{save_full_screen}
\title{Create and save all raw data from the screen, before hit/sensitive assignment.}
\usage{
save_full_screen(tidied_screen, ratio_marg = 1, ...)
}
\arguments{
\item{tidied_screen}{a tibble containing raw data from a dye screen, both protein and buffer, e.g. as output by the \code{tidy_dye_screen()} function.}

\item{ratio_marg}{a number, giving the ratio between the total size of a facet's plot panel and the their facet panel, inclusive of borders, axis ticks, etc. Ideally this could be reliably extracted from the plot object itself in a future version of this package. Is greater than one.}

\item{...}{additional arguments, passed to make_figure_title() and save_stacked_plots(). See make_figure_title() and save_stacked_plots() for possible options, which include assigning a directory to which to write the plot object created, and requesting the plot be saved using a specific  the graphics device  (default is "pdf" but all extentions compatible with ggsave() can be used).}
}
\value{
if assigned to an object, returns a ggproto object. Always saves a figure to the current working directory, or the directory specified through arguments to ...
}
\description{
Creates a plot object, and saves a figure, containing raw data for all screened dyes in all measured fluorescent channels. Results are divided into two subplots, containing dyes which are declared hits, and sensitives. Each dye is represented in a single row, with each column containing raw data collected for that dye in a given fluorescent channel. Both protein and buffer data are displayed.
}
\details{
This plot is closely related to the one created by the \code{save_stacked_screen()} function, but differs in that dyes are not yet divided into "hit", "sensitive", and "none" categories. It is often used during the process of creating those assignments, and as an output of the default dye screening workflow, this plot serves as an unbiased representation of screening data, so that screens can be re-assessed with fresh eyes after hits have been assigned if desired.

This function is often accessed through it's wrapper function, save_dye_screen_figs(), though it can be used identically on it's own. It is the third of the four standard raw data plots created in a default dye-screening workflow, where the other four plots are (1) all raw data, with no assignments displayed. Used for calling hits, as created by the \code{save_full_screen()} function; (2) all raw data, with distinct sub-plots containing dyes given each assignment--hit, sensitive, or none--as created by the \code{save_stacked_screen()} function; (3) raw data for just dyes assigned to "hit" and "sensitive" categories, with data collected for each dye displayed in a separate row, and data for the raw data collected for that dye in each fluorescent channel displayed in a separate colummn; (4) the hit and/or sensitive dyes, in specific channels, carried through model fitting and tma extraction.
}
