% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dye_screen_model_plots.R
\name{save_model_figure}
\alias{save_model_figure}
\title{Save a figure displaying raw data, fits, and tmas. Use reasonable dimensions.}
\usage{
save_model_figure(
  tidied_for_tmas,
  by_var_fit,
  tidied_screen,
  hits_by_channel,
  extract_figure_title = TRUE,
  extract_layout_from_tidied = TRUE,
  extract_tmas_from_by_var_fit = TRUE,
  layout = NULL,
  model_tmas = NULL,
  .well_col_layout = "variable",
  ...
)
}
\arguments{
\item{tidied_for_tmas}{a grouped, nested tibble, as returned by \code{tidy_for_tmas}}

\item{by_var_fit}{a tibble, as output by \code{add_model_fits}.}

\item{tidied_screen}{a tibble, as returned by \code{tidy_dye_screen()}}

\item{hits_by_channel}{a tibble containing hit assignments, as expected by \code{pull_assigned()}}

\item{extract_figure_title}{a boolean; if TRUE, both saved name and plot title are extracted from the given data via \code{make_figure_title()}}

\item{extract_layout_from_tidied}{a boolean; if TRUE, the experimental layout is extracted from \code{tidied_screen} via \code{extract_plate_layout}, and the unique-identifier column of the output layout is named ".var"}

\item{extract_tmas_from_by_var_fit}{a boolean; if TRUE, the tmas applied to the plots are extracted directly from \code{by_var_fit}. This is usually done in a single operation, and it's not clear exactly why one would set this to FALSE unless avoid that additional operation was important, and the model tma values had already been calculated and stored elsewhere.}

\item{layout}{if extract_layout_from_tidied is FALSE, a tibble containig a layout, as returned by \code{read_plate_layout} or \code{extract_plate_layout}. Gets passed to \code{tidy_for_plot()} in this function.}

\item{model_tmas}{if extract_tmas_from_by_var_fit is FALSE, a tibble containing the extracted model tmas, formatted to match the output of \code{extract_model_tmas()}.}

\item{.well_col_layout}{a string, giving the name of the column containing well names in \code{tidied_for_tmas}, which is renamed ".var" if extract_layout_from_tidied = TRUE.}

\item{...}{additional named parameters, passed to the functions: \code{extract_plate_layout()}, \code{join_with_preds()}, \code{tidied_for_plot()}, \code{plot_analyzed_hits()}, \code{force_panel_sizing()}, and \code{save_stacked_plots()}}
}
\value{
if assigned, the final figure object. regardless of assignment, a saved figure, with names, directories, and file types as dictated by \code{save_stacked_plots()}. Uses \code{save_stacked_plots()} defaults, unless these arguments are overwritten by passing new named values for them to this function.
}
\description{
A wrapper function for adding facets, calculating dimenstions, and saving with a consistent name. In a better version of this function, would likely be refactored alongside \code{plot_with_preds()} and \code{plot_analyzed_hits()}. Calls on \code{save_stacked_plot()} to do so.
}
