% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sgolay}
\alias{sgolay}
\title{Wrapper for Savistky-Golay filtering}
\usage{
sgolay(x, p, n, m)
}
\arguments{
\item{x}{a numeric vector to be savitsky-golay filtered}

\item{p}{the filter order to use. Defaults to 5. Must be odd and >2.}

\item{n}{the number of measurements in the filter window. Defaults to 13.}

\item{m}{the filter order. 0: smoothing only; 1: first derivative, 2: second derivative, etc.}
}
\value{
a numeric vector, with a savistky-golay filter applied.
}
\description{
A closure which sets defaults frequently appropriate for DSF for p (filter order) and n (filter length). Allows the filter to be called with the simplier syntax later. e.g. sgolay(value, 1) returns the first derivative of the vector 'value'. Formals for sgolay and be reset after calculations performed on user data to determine ideal p and n. This behavior is not implemented yet.
}
\details{
Savistky-Golay filtering is used for all smoothing and derivative calculations in dsfworld.
}
