% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_tmas.R
\name{sort_tmas}
\alias{sort_tmas}
\title{Sort tmas returned from models into tma1, tma2, etc}
\usage{
sort_tmas(df, sort_type = "low_high", close_val = 0, .tma_col = "tma")
}
\arguments{
\item{df}{df of tmas}

\item{sort_type}{sorting method. defaults to "low_high". Also can be "closest_to"}

\item{close_val}{if closest to, the value to rank tmas by proximity to.}

\item{.tma_col}{name of column containing tma data. defaults to tma.}
}
\value{
the df, with a "rank" column
}
\description{
Sort tmas returned from models into tma1, tma2, etc
}
