% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dye_screen.R
\name{theme_screen_austere}
\alias{theme_screen_austere}
\title{Default dye screen plot theme}
\usage{
theme_screen_austere(...)
}
\arguments{
\item{...}{additional theme arguments which overwrite defaults within theme_bw()}
}
\value{
a plot theme which can be applied to any ggplot2 object with theme_screen_austere()
}
\description{
ggplot2 theme used in the dye screen plots.
}
