% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_dye_screen.R
\name{tidy_dye_screen}
\alias{tidy_dye_screen}
\title{Tidy dye screens, combine buffer and protein data}
\usage{
tidy_dye_screen(
  .raw_data,
  .raw_layout,
  .buffer_data,
  .buffer_layout,
  .protein_name,
  .exp_num,
  .buffer_used,
  ...
)
}
\arguments{
\item{.raw_data}{a long-form tibble containing a single plate of raw DSF data from a dye screen. Must include columns containing information on each of the following things:
\itemize{
\item \strong{well}, or some other individual reaction-identifying information. This column is coerced to character-type in the output of this function.
\item \strong{Temperature}, giving the Temperature of each RFU measurement. This column is coerced to numeric-type in the output of this function.
\item \strong{Wavelength of measurement} A column giving information about the wavelengths at which measurements were made, e.g. "channel_f". This column must be a factor; it is recommended that the factor levels reflect the wavelengths measured by each channel. For example, "channel_f" may be a factor column, with levels of "FAM", "JOE", "TAMRA", "ROX", "Cy5", "Cy5.5". This column is added automatically by read_qtower, but may need to be generated manually for other types of inputs.

\item \strong{Raw RFU value}, giving the raw values measured. This column is coerced to numeric-type in the output of this function. A by-variable normalized version of this column is also present in the returned tibble.
}}

\item{.raw_layout}{a long-form tibble containing plate layout information for the supplied raw data. Must include columns containing information on each of the following things:
#' \itemize{
\item \strong{well}, or some other individual reaction-identifying information which matches the identifying column in the supplied data.nThis column is coerced to character-type in the output of this function.
\item \strong{dye}, giving a unique identifying name for the dye that was present in a well. This column is coerced to character-type in the output of this function.
\item \strong{dye concentration}, a column giving the concentration (in uM) at which the dye was tested. This column is coerced to numeric-type in the output of this function. uM units are not precisely necessary, but the default name of this column for processing is "dye_conc_uM", so non-uM units may lead to misleading column labeling if default column names are used.
}}

\item{.buffer_data}{the same as .raw_data, but data is from a no-protein control.}

\item{.buffer_layout}{the same as .raw_data, but data is from a no-protein control.}

\item{.protein_name}{the name of the protein screened.}

\item{.exp_num}{a unique identifier which can assist in tracing this output data back to a specific experiment or notebook entry.}

\item{.buffer_used}{a single string containing brief information on the buffer used in this screen.}

\item{...}{additional arguments passed to label_dye_screen(), and label_data(), which is called within label_dye_screen().}
}
\value{
a tibble, containing data for a labeled dye screen, containing both protein and buffer data.
\itemize{
\item \strong{variable}, a character column holding unique identifiers for each trace.
\item \strong{Temperature}, a numeric column holding Temperatures at which RFU measurements were made.
\item \strong{value}, a numeric column holding raw RFU values measured.
\item \strong{value_group_norm}, a numeric column holding raw RFU values, in which buffer and protein data for a particular dye, at a particular concentration, in a particular channel, collectively have been normalized to a 0 to 1 scale.
\item \strong{value_norm}, a numeric column holding raw RFU values, scales 0 to 1 for each individual trace.
\item \strong{dye}, a character column specifying the dye tested in each well.
\item \strong{dye_conc_uM}, a character column specifying the concentration at which a dye was tested.
\item \strong{channel_f}, a character column specifying the channel in which the raw values were measured.
\item \strong{type}, a character column specifying the type of moleceule screened-- "protein" or "buffer". Each output may contain only one value for "type". For plates in which buffer and protein were both screened, layouts which mask the wells of the opposite type should be used, to create separate outputs for each type of data.
\item \strong{identity}, a character column holding specific identities for the protein and buffer screened, e.g. "lysozyme" and "20mM_HEPES_pH7p2_200mM_NaCl_1mM_TCEP".
\item \strong{exp_num}, a character column holding information which can be used to trace the experiment back to a specific notebook entry or other form of documentation.
}
}
\description{
Combine protein and buffer dye screening data, and output a dye screening dataframe, tidied for downstream analysis.
}
