% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_for_tmas.R
\name{tidy_for_tmas}
\alias{tidy_for_tmas}
\title{Tidy raw dsf dataframe for tma calculations}
\usage{
tidy_for_tmas(
  df,
  unique_col = "well",
  value_col = "value",
  temp_col = "Temperature",
  keep_cols = c("well", "Temperature", "value", "value_norm")
)
}
\arguments{
\item{df}{tibble containing long-form raw dsf data}

\item{unique_col}{name of the column containing a unique identifier for each dataset. Defaults to "well".}

\item{value_col}{name of the column containing raw rfu values. Defaults to "value".}

\item{temp_col}{name of the column containing temperatures. Defaults to "Temperature"}

\item{keep_cols}{name of all column to keep. Defaults to minimum: "well", "Temperature", "value", and "value_norm"}
}
\value{
a nested tibble containig tidied data for downstream tma analyses. Column names are .var (for unique identifier), and data (for the nested data corresponding to each value of .var). Each nested tibble contains numeric columns: Temperature, Temperature_norm, value, value_norm, and drfu_norm.
}
\description{
Also adds 0 - 1 range normalizations of temperature and raw data, and the first derivative of the normalizsed raw value.
}
