% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_fits.R
\name{add_component_predictions}
\alias{add_component_predictions}
\title{Add predictions for each component to the dataframe}
\usage{
add_component_predictions(
  tidied,
  data,
  temp_col = "Temperature",
  rescale_temp_col = TRUE
)
}
\arguments{
\item{tidied}{a tibble; the tidied model outputs}

\item{data}{the column containing nested data}

\item{temp_col}{name of the column containing temperature in data}

\item{rescale_temp_col}{normalize the temperature column?}
}
\value{
data df, with predicts added
}
\description{
Add predictions for each component to the dataframe
}
