% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model_pars.R
\name{add_estimates}
\alias{add_estimates}
\title{Add locations of estimated transitions to dataframe}
\usage{
add_estimates(
  df,
  .est_peak_col = "drfu_norm",
  .est_init_col = "value_norm",
  .norm_raw_x = "Temperature_norm",
  .norm_raw_y = "value_norm"
)
}
\arguments{
\item{df}{nested tibble of dsf data}

\item{.est_peak_col}{name of column from which to estimate peaks defaults to "drfu_norm"}

\item{.est_init_col}{name of column from which to estimate inital rfu. defaults to "value_norm"}

\item{.norm_raw_x}{name of column containing normalized temperature. defaults to "Temperature_norm"}

\item{.norm_raw_y}{name of column containing normalized raw RFU. defaults to "value_norm"}
}
\value{
df, with a new column with estimations
}
\description{
Add locations of estimated transitions to dataframe
}
