% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_fits.R
\name{add_model_fits}
\alias{add_model_fits}
\title{Extract apparent tms from model component fits}
\usage{
add_model_fits(
  by_var,
  which_models,
  dsfworld_models = TRUE,
  model_pars = NULL,
  model_formula = NULL,
  model_lower_bound = NULL,
  ...
)
}
\arguments{
\item{by_var}{by var df}

\item{which_models}{models to fit}

\item{dsfworld_models}{use dsfworld models?}

\item{model_pars}{defaults to NULL, user supplies if using their own model}

\item{model_formula}{defaults to NULL, user supplies if using their own model}

\item{model_lower_bound}{defaults to NULL, user supplies if using their own model}

\item{...}{passed to the model fitting}
}
\value{
a df containing model tmas
}
\description{
Extract apparent tms from model component fits
}
