% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_model_preds.R
\name{annotate_tidied_components}
\alias{annotate_tidied_components}
\title{Annotate tidied components}
\usage{
annotate_tidied_components(tidied)
}
\arguments{
\item{tidied}{a tibble resulting from broom::tidy of a dsfworld model fit.}
}
\value{
the same tibble, with columns added: "term_type", and "component", to facilitate reconstruction and formulas to predict components.
}
\description{
A helper function for add_component_predictions
}
