% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_model_tmas.R
\name{extract_component_tma}
\alias{extract_component_tma}
\title{Extract component tmas}
\usage{
extract_component_tma(
  data,
  component_names = c("sigmoid_1", "sigmoid_2", "sigmoid_3"),
  ...
)
}
\arguments{
\item{data}{a tibble, as returned by add_model_preds().}

\item{component_names}{the names of the components from which to extract Tmas. Defaults to c("sigmoid_1", "sigmoid_2", "sigmoid_3"). Components listed but not found in data are silently ignored.}

\item{...}{additional arguments, passed to get_drfu_tmas. Also serves the purpose of allowing irrelevant arguments passed from upstream functions using ... to be ignored.}
}
\value{
a tibble containing two columns, "which_value", a character defining which of the input component_names the Tma corresponds to, and "tma", the tma value determined using get_drfu_tmas. Note that the tmas will be for the x vector to which the models were fit, which in the standard workflow, is "Temperature_norm". Conversion to real temperatures currently is handled outside of the fitting workflow.
}
\description{
A helper function for add_model_tmas. Extracts apparent melting temperatures from individually-predicted sigmoids from model fits. Uses maximum of dRFU to do this.
}
