% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_model_preds.R
\name{get_component_exprs}
\alias{get_component_exprs}
\title{Extract indiviidual component expressions from  model formulas}
\usage{
get_component_exprs(component_type)
}
\arguments{
\item{component_type}{the type of component to extract expressions from; "sigmoid", or "initial"}
}
\value{
a quoted fomrula for an individual component, which is evaluated in downstream functions to reconstruct individual sigmoid components from model fits.
}
\description{
Extract indiviidual component expressions from  model formulas
}
