% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_start_pars.R
\name{get_start_pars}
\alias{get_start_pars}
\title{Select and format estimates for model starting parameters}
\usage{
get_start_pars(
  estimates,
  which_pars,
  pars_defaults = list(Asym1 = 1, xmid1 = 0.4, scal1 = 0.03, d1 = -0.5, Asym2 = 0.5,
    xmid2 = 0.6, scal2 = 0.03, d2 = -1, Asym3 = 0.3, xmid3 = 0.2, scal3 = 0.03, d3 =
    -1.5, id_d1 = 0.2, id_b1 = -5),
  par_order = c("Asym1", "xmid1", "scal1", "d1", "Asym2", "xmid2", "scal2", "d2",
    "Asym3", "xmid3", "scal3", "d3", "id_d1", "id_b1"),
  ...
)
}
\arguments{
\item{estimates}{a tibble containing all estimates}

\item{which_pars}{names of the parameters to get starting estimates for}

\item{pars_defaults}{default values for all parameters}

\item{par_order}{order in which nlsLM will expect the parameters in the downstream funtions}

\item{...}{do this}
}
\value{
parameter list for passing to fitting function (uses nlsLM)
}
\description{
Select and format estimates for model starting parameters
}
