% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_model_preds.R
\name{predict_component}
\alias{predict_component}
\title{Predict y values for an individual component}
\usage{
predict_component(
  df,
  temp_vec,
  .component_type_col = "component_type",
  .est_col = "estimate",
  .term_type_col = "term_type"
)
}
\arguments{
\item{df}{data, after drop_and_pivot, to which the predicted component values will be added.}

\item{temp_vec}{numeric vector containing the temperature}

\item{.component_type_col}{the name of the column in the df argument, containing the componet type to be reconstructed.}

\item{.est_col}{a string, giving the name of the column containing estimates values.}

\item{.term_type_col}{a string, giving the name of the column containing the term type values.}
}
\value{
a tibble containing two columns, x and y, containing the input temperature vector and the predicted y values for the predicted component.
}
\description{
A helper function for get_component_predictions. Given a qutoed formula and x data, predicts the values of he individual components.
}
